+++ tests/Raid/RaidZeroTest.php
@@ -1,11 +1,11 @@
 <?php
-namespace kevinquinnyo\Raid\Test;
+namespace kevinquinnyo\Raid\Test\Raid;
 
 use \PHPUnit\Framework\TestCase;
 use \kevinquinnyo\Raid\Drive;
-use \kevinquinnyo\Raid\Raid\RaidOne;
+use \kevinquinnyo\Raid\Raid\RaidZero;
 
-class RaidOneTest extends TestCase
+class RaidZeroTest extends TestCase
 {
     public function testGetCapacity()
     {
@@ -13,10 +13,11 @@ class RaidOneTest extends TestCase
             new Drive(1024, 'ssd', 1),
             new Drive(1024, 'ssd', 2),
         ];
-        $raidOne = new RaidOne($drives);
-        $this->assertSame(1024, $raidOne->getCapacity());
-        $this->assertSame('1 KB', $raidOne->getCapacity(['human' => true]));
+        $raidZero = new RaidZero($drives);
+        $this->assertSame(2048, $raidZero->getCapacity());
+        $this->assertSame('2 KB', $raidZero->getCapacity(['human' => true]));
     }
+
     public function testGetCapacityWithHotSpares()
     {
         $drives = [
@@ -26,12 +27,13 @@ class RaidOneTest extends TestCase
             new Drive(1024, 'ssd', 4, ['hotSpare' => true]),
             new Drive(1024, 'ssd', 5, ['hotSpare' => true]),
         ];
-        $raidOne = new RaidOne($drives);
-        $this->assertSame(1024, $raidOne->getCapacity());
+        $raidZero = new RaidZero($drives);
+        $this->assertSame(3072, $raidZero->getCapacity());
     }
+
     public function testGetLevel()
     {
-        $raidOne = new RaidOne();
-        $this->assertSame(1, $raidOne->getLevel());
+        $raidZero = new RaidZero();
+        $this->assertSame(0, $raidZero->getLevel());
     }
 }
